+++ UnitTests/BraintreeThreeDSecureTests/BraintreeThreeDSecureTests-Bridging-Header.h
@@ -8,3 +8,4 @@
 #import <BraintreeThreeDSecure/BTThreeDSecureV1BrowserSwitchHelper.h>
 #import <BraintreeThreeDSecure/BTThreeDSecureV2UICustomization_Internal.h>
 #import <BraintreeThreeDSecure/BTThreeDSecureV2BaseCustomization_Internal.h>
+#import <BraintreeCore/Braintree-Version.h>